pyproject.toml
[project]
name = "onchaindecoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "uvicorn>=0.41.0",
    "web3>=7.14.1",
]
